+++ src/package/tyhUi/src/index.js
@@ -38,7 +38,6 @@ import TyhOption from '../packages/option'
 import TyhTagging from '../packages/tagging'
 import { Message } from '../packages/message'
 import { Notification } from '../packages/notification'
-console.log(Notification)
 
 const components = [
   TyhButton,
@@ -88,7 +87,7 @@ const install = app => {
 }
 
 const tyhUi = {
-  version: "3.3.0",
+  version: "3.3.1",
   install
 }